dom/canvas/test/test_imagebitmap_structuredclone_iframe.html

<!DOCTYPE HTML>
<title>Test ImageBitmap : StructuredClone between main window and iframe</title>
<meta charset="utf-8">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<div id="content"></div>
<script src="imagebitmap_structuredclone_utils.js"></script>
<script type="text/javascript">

function runTests() {
  window.onmessage = function(event) {
    if (event.data.type == "status") {
      ok(event.data.status, event.data.msg);
    } else if (event.data.type == "finish") {
      SimpleTest.finish();
    } else if (event.data.type == "bitmap1") {
      compareImageBitmapWithImageElement(event.data.bitmap, gImage1);
    } else if (event.data.type == "bitmap2") {
      compareImageBitmapWithImageElement(event.data.bitmap, gImage2);
    } else if (event.data.type == "bitmap3") {
      compareImageBitmapWithImageData(event.data.bitmap, gImageData, "Check preserving alpha");
    }
  }

  var div = document.getElementById("content");
  ok(div, "Parent exists");

  var ifr = document.createElement("iframe");
  ifr.addEventListener("load", iframeLoaded, false);
  ifr.setAttribute('src', "imagebitmap_structuredclone_iframe.html");
  div.appendChild(ifr);

  function iframeLoaded() {
    prepareImageBitmaps().then(function(){
      ifr.contentWindow.postMessage({"bitmap1":gImageBitmap1, "bitmap2":gImageBitmap2, "bitmap3":gImageBitmap3}, "*");
    });
  }
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTests);

</script>
</body>
